+++ deployment/terraform/dataspace/outputs.tf
@@ -1,10 +1,18 @@
+output "connector_name" {
+  value = local.connector_name
+}
+
+output "key_vault" {
+  value = azurerm_key_vault.registrationservice.name
+}
+
 output "app_insights_connection_string" {
   value     = azurerm_application_insights.dataspace.connection_string
   sensitive = true
 }
 
 output "registration_service_url" {
-  value = "http://${azurerm_container_group.registration-service.fqdn}:${local.edc_default_port}"
+  value = local.registration_service_url
 }
 
 output "dataspace_did_host" {